ANOVA Tables

anova_tables(
  x,
  effects = "fixed",
  contrast = NULL,
  at = NULL,
  standardized = TRUE,
  unstandardized = TRUE,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = NULL,
  pbkrtest.limit = NULL,
  lmerTest.limit = NULL,
  digits = 3,
  id_col = "Subject",
  print = TRUE
)

Arguments

x
an lmer model object

effects
"fixed" or "all". default is "fixed" to reduce computation time

contrast
The factor(s) at which to compare levels at

at
Additional interacting factor(s) to compare the effect of contrast at

standardized
Logical, indicating whether or not to print standardized estimates. Standardized estimates are based on "refit" of the model on standardized data but it will not standardize categorical predictors. Defualt is TRUE.

unstandardized
Logical, indicating whether or not to print unstandardized estimates. Default is TRUE.

ci
Confidence Interval (CI) level. Default to 0.95 (95

ci_methodDocumention based on ?parameters::parameters. Method for computing degrees of freedom for confidence intervals (CI) and the related p-values. Allowed are following options (which vary depending on the model class): "residual", "normal", "likelihood", "satterthwaite", "kenward", "wald", "profile", "boot", "uniroot", "ml1", "betwithin", "hdi", "quantile", "ci", "eti", "si", "bci", or "bcai". See section Confidence intervals and approximation of degrees of freedom in model_parameters() for further details. When ci_method=NULL, in most cases "wald" is used then.

bootstrapDocumention based on ?parameters::parameters. Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterationsDocumention based on ?parameters::parameters. The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

pbkrtest.limitOptional parameter that can be set to help calculate dfs. If you need to use this a warning message will appear in the console telling you what to set this at.

lmerTest.limitOptional parameter that can be set to help calculate dfs. If you need to use this a warning message will appear in the console telling you what to set this at.

digitsHow many decimal places to round to? Default is 3.

id_colThe column containing subject ids. Default is "Subject"

printCreate a knitr table for displaying as html table? (default = TRUE)